refactor daemon folder
