+++ src/main/python/daemon/git_attributes.txt
